requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
pandas
requests
bs4
selenium
beautifulsoup4
selenium
python-dotenv